backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bcrypt
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dataclasses-json
distro
et_xmlfile
fastapi
frozenlist
greenlet
h11
httpcore
httpx
idna
invoke
Jinja2
jiter
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
lxml
MarkupSafe
marshmallow
multidict
mypy_extensions
numpy
openai
openpyxl
orjson
packaging
pandas
paramiko
passlib
pgvector
propcache
psycopg2-binary
pycparser
pydantic
pydantic_core
PyJWT
PyNaCl
python-dateutil
python-docx
python-dotenv
python-multipart
PyYAML
regex
requests
requests-toolbelt
six
sniffio
SQLAlchemy
starlette
tenacity
tiktoken
tqdm
typing-inspect
typing_extensions
urllib3
uvicorn
yarl
zhipuai
